requirements.txt
Flask
line-bot-sdk
pdfplumber
gunicorn